+++ src/internal/common/messageprotocol/rabbit/channels/accountid/accountid.go
@@ -8,11 +8,11 @@ import (
 
 type Msg = batch.Msg[account.AccountIdentifier]
 
-func WriteBatch(clientID int, queryID uint8, senderID uint32, seq uint32, ids []account.AccountIdentifier) []byte {
+func WriteBatch(clientID int, queryID uint8, senderID uint8, seq uint64, ids []account.AccountIdentifier) []byte {
 	return batch.Write(clientID, queryID, senderID, seq, ids, records.AccountIdentifierCodec)
 }
 
-func WriteEOF(clientID int, queryID uint8, senderID uint32, seq uint32, total uint32) []byte {
+func WriteEOF(clientID int, queryID uint8, senderID uint8, seq uint64, total uint32) []byte {
 	return batch.WriteEOF(clientID, queryID, senderID, seq, total)
 }
