+++ src/internal/common/messageprotocol/rabbit/channels/qualifiedaccount/qualifiedaccount.go
@@ -25,11 +25,11 @@ func NewBatchBuilder(maxCount, maxBytes int) *batch.Builder[QualifiedAccount] {
 	return batch.NewBuilder(maxCount, maxBytes, codec)
 }
 
-func WriteBatch(clientID int, queryID uint8, senderID uint32, seq uint32, recs []QualifiedAccount) []byte {
+func WriteBatch(clientID int, queryID uint8, senderID uint8, seq uint64, recs []QualifiedAccount) []byte {
 	return batch.Write(clientID, queryID, senderID, seq, recs, codec)
 }
 
-func WriteEOF(clientID int, queryID uint8, senderID uint32, seq uint32, total uint32) []byte {
+func WriteEOF(clientID int, queryID uint8, senderID uint8, seq uint64, total uint32) []byte {
 	return batch.WriteEOF(clientID, queryID, senderID, seq, total)
 }
